processing/requirements.txt
# Real-Time Data Pipeline – Python dependencies

# Kafka
kafka-python==2.0.2

# PySpark + Delta Lake (Spark consumer)
pyspark==3.5.1
delta-spark==3.1.0

# Message queue integrations
pika==1.3.2           # RabbitMQ
azure-servicebus==7.12.1  # Azure Service Bus

# Utilities
python-dotenv==1.0.1
